Make menu_color setting permanent with a config
In the previous commit, the menu color setting wasn't saved, and after rebooting it was lost, resulting a recovery which had always blue color after rebooting, even if the user wanted orange or green. So, he always had to press the button corresponding to the color, after every reboot if he wanted green or orange. In this commit, that setting is saved in /sdcard/clockworkmod/cwm_config which assigns a value to each config and which is read before displaying the color, so that if contains color=1/2/3, it will use the color corresponding to the value, even after rebooting